qwen-vl-service/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.49.0
tokenizers>=0.15.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Qwen VL specific
qwen-vl-utils[decord]==0.0.8

# Vision/Video processing
opencv-python>=4.5.0
pillow>=9.0.0
decord>=0.6.0

# Service framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Optimization libraries
onnx>=1.14.0
onnxruntime>=1.15.0

# Mobile deployment (optional)
coremltools>=7.0; sys_platform == "darwin"
tensorflow>=2.13.0

# Monitoring and logging
prometheus-client>=0.16.0
redis>=4.5.0

# Development tools
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Data processing
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
